% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_reference_ei.R
\name{child_reference_EI}
\alias{child_reference_EI}
\title{Energy Intake Matrix}
\usage{
child_reference_EI(age, sex, FM, FFM, days)
}
\arguments{
\item{age}{(vector) Age of individual (yrs)}

\item{sex}{(vector) Sex either \code{"female"} or \code{"male"}}

\item{FM}{(vector) Fat Mass at Baseline}

\item{FFM}{(vector) Fat Free Mass at Baseline

\strong{ Optional }}

\item{days}{(numeric) Days to run the model.}
}
\description{
Estimates weight given age, sex, fat mass, and fat free mass,
}
\examples{
#One child
child_reference_EI(6, "male", 2, 4, 10)

#Several children
child_reference_EI(sample(6:12, 10, replace = TRUE), 
                   sample(c("male","female"), 10, replace = TRUE), 
                   sample(2:10, 10, replace = TRUE), 
                   sample(2:10, 10, replace = TRUE),
                   365)

}
\references{
Hall, K. D., Butte, N. F., Swinburn, B. A., & Chow, C. C. (2013). 
\emph{Dynamics of childhood growth and obesity: development and validation of a 
quantitative mathematical model}. The Lancet Diabetes & Endocrinology, 1(2), 97-105.
}
\seealso{
\code{\link{model_plot}} for plotting the results and 
\code{\link{model_mean}} for aggregate data estimation.
}
\author{
Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

Dalia Camacho-García-Formentí \email{daliaf172@gmail.com}
}
\keyword{internal}
