% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_reference_FFMandFM.R
\name{child_reference_FFMandFM}
\alias{child_reference_FFMandFM}
\title{FFM and FM reference}
\usage{
child_reference_FFMandFM(age, sex)
}
\arguments{
\item{age}{(vector) Age of individual (yrs)}

\item{sex}{(vector) Sex either \code{"female"} or \code{"male"}

\strong{ Optional }}
}
\description{
Estimates FFM and FM reference given age, sex.
}
\examples{
#One child
child_reference_FFMandFM(6, "male")

#Several children
child_reference_FFMandFM(sample(6:12, 10, replace = TRUE), 
                   sample(c("male","female"), 10, replace = TRUE))

}
\references{
Hall, K. D., Butte, N. F., Swinburn, B. A., & Chow, C. C. (2013). 
\emph{Dynamics of childhood growth and obesity: development and validation of a 
quantitative mathematical model}. The Lancet Diabetes & Endocrinology, 1(2), 97-105.
}
\seealso{
\code{\link{model_plot}} for plotting the results and 
\code{\link{model_mean}} for aggregate data estimation.
}
\author{
Rodrigo Zepeda-Tello \email{rzepeda17@gmail.com}

Dalia Camacho-García-Formentí \email{daliaf172@gmail.com}
}
\keyword{internal}
